Fix: don't clear display for each drawing
When we attempt to render more than 2 parts, we couldn't hold previous
parts because each time 'Draw::DrawAt' executed, the display was cleared.
So Changed not to clear the display for each time.

BY DOING THIS, YOU SHOULD FILL WHOLE CANBUS(where any letters could
appear in other parts) OR SURPLUS PARTS WILL BE LEFT WHEN YOU OVERWRITE IT.